feat(opencode-agent): render phase reports as sections of one comment
First step of the one-command-one-reply change: `renderStatus` gains
`sections`, rendering each phase's report oldest-first with every one but the
newest folded away, and the run's own progress table collapses into a `Run
detail` disclosure below them.

Sections pushed the file past max-lines, so the run's self-description —
progress table, job/branch and budget lines — moves to `run-detail.ts`,
leaving `status-comment.ts` holding what a reply is made of.

Additive: nothing posts sections yet, and the live channel is untouched.
